Multimedia-System-Design/Video_Summary/requirements.txt
SoundFile==0.10.3.post1
ffmpeg_python==0.2.0
scikit_image==0.17.2
imageio==2.5.0
scipy==1.1.0
keras_segmentation==0.3.0
opencv_python==4.4.0.46
numpy==1.19.5
matplotlib==3.3.3
simpleaudio==1.0.4
sounddevice==0.4.1
ffmpeg==1.4
progressbar33==2.4
scikit_learn==0.24.2
skimage==0.0
